adding users authentication

+++ backend/app/Http/Controllers/TodosController.php
@@ -14,7 +14,7 @@ class TodosController extends Controller
      */
     public function index()
     {
-        return Todo::all();
+        return Todo::where('user_id', auth()->user()->id)->get();
     }
 
     /**
@@ -27,10 +27,14 @@ class TodosController extends Controller
     {
         $data = $request->validate([
             'title' => 'required|string',
-            'completed' => 'required|boolean'
+            'completed' => 'required|boolean',
         ]);
 
-        $todo = Todo::create($data);
+        $todo = Todo::create([
+            'user_id' => auth()->user()->id,
+            'title' => $request->title,
+            'completed' => $request->completed,
+        ]);
 
         return response($todo, 201);
     }
@@ -44,9 +48,13 @@ class TodosController extends Controller
      */
     public function update(Request $request, Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $data = $request->validate([
             'title' => 'required|string',
-            'completed' => 'required|boolean'
+            'completed' => 'required|boolean',
         ]);
 
         $todo->update($data);
@@ -57,12 +65,12 @@ class TodosController extends Controller
     public function updateAll(Request $request)
     {
         $data = $request->validate([
-            'completed' => 'required|boolean'
+            'completed' => 'required|boolean',
         ]);
 
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
 
-        return response("Updated", 200);
+        return response()->json('Updated', 200);
     }
 
     /**
@@ -73,19 +81,40 @@ class TodosController extends Controller
      */
     public function destroy(Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $todo->delete();
 
-        return response("Deleted todo item", 200);
+        return response()->json('Deleted todo item', 200);
     }
 
     public function destroyCompleted(Request $request)
     {
+        // [6,9] todo ids we are passing in and want to delete
+        // [5,6,9] all of the users todo ids
+
+        $todosToDelete = $request->todos;
+
+        $userTodoIds = auth()->user()->todos->map(function ($todo) {
+            return $todo->id;
+        });
+
+        $valid = collect($todosToDelete)->every(function ($value, $key) use ($userTodoIds) {
+            return $userTodoIds->contains($value);
+        });
+
+        if (!$valid) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $request->validate([
-            'todos' => 'required|array'
+            'todos' => 'required|array',
         ]);
 
         Todo::destroy($request->todos);
 
-        return response("Deleted", 200);
+        return response()->json('Deleted', 200);
     }
 }
